sandbox/my_special_project/requirements.txt
asgiref
Django
django-admin
django-excel-response2
django-six
excel-base
isoweek
python-dateutil
python-dotenv
pytz-deprecation-shim
screen
six
sqlparse
TimeConvert
tzdata
tzlocal
xlwt
django-cors-headers
sqlparse
whitenoise
gunicorn
djangorestframework 
django-filter 
django-restframework
djangorestframework-simplejwt
Pillow 
psycopg2-binary 
twilio 
django-storages 
google-cloud-storage 
celery 
redis 
